+++ packages/visual-editor/src/components/_componentCategories.ts
@@ -7,25 +7,40 @@ import {
   CoreInfoSection,
   CoreInfoSectionProps,
 } from "./pageSections/CoreInfoSection.tsx";
+import {
+  EventSection,
+  EventSectionProps,
+} from "./pageSections/EventSection.tsx";
 import { HeroSection, HeroSectionProps } from "./pageSections/HeroSection.tsx";
+import {
+  InsightSectionProps,
+  InsightSection,
+} from "./pageSections/InsightSection.tsx";
 import {
   NearbyLocationsSection,
   NearbyLocationsSectionProps,
 } from "./pageSections/NearbyLocations.tsx";
-import { PromoSection, PromoSectionProps } from "./pageSections/Promo.tsx";
+import {
+  ProductSection,
+  ProductSectionProps,
+} from "./pageSections/ProductSection.tsx";
+import {
+  PromoSection,
+  PromoSectionProps,
+} from "./pageSections/PromoSection.tsx";
 import {
   PhotoGallerySection,
   PhotoGallerySectionProps,
 } from "./pageSections/PhotoGallerySection.tsx";
-import { FAQsSection, FAQsSectionProps } from "./pageSections/FAQs.tsx";
+import { FAQSection, FAQSectionProps } from "./pageSections/FAQsSection.tsx";
 import {
-  SectionContainer,
-  SectionContainerProps,
-} from "./pageSections/SectionContainer.tsx";
+  TestimonialSection,
+  TestimonialSectionProps,
+} from "./pageSections/TestimonialSection.tsx";
+import { TeamSection, TeamSectionProps } from "./pageSections/TeamSection.tsx";
 import { Collection, CollectionProps } from "./layoutBlocks/Collection.tsx";
 import { Flex, FlexProps } from "./layoutBlocks/Flex.tsx";
 import { Grid, GridProps } from "./layoutBlocks/Grid.tsx";
-import { ProductCard, ProductCardProps } from "./cards/ProductCard.tsx";
 import { Address, AddressProps } from "./contentBlocks/Address.tsx";
 import { BodyText, BodyTextProps } from "./contentBlocks/BodyText.tsx";
 import { CTAWrapper, CTAWrapperProps } from "./contentBlocks/CtaWrapper.tsx";
@@ -48,23 +63,33 @@ import { Header, HeaderProps } from "./Header.tsx";
 import { Footer, FooterProps } from "./Footer.tsx";
 import { Directory, DirectoryProps } from "./Directory.tsx";
 import { EventCardProps, EventCard } from "./cards/EventCard.tsx";
-import { InsightCardProps, InsightCard } from "./cards/InsightCard.tsx";
+import { ProductCardProps, ProductCard } from "./cards/ProductCard.tsx";
 import { PersonCardProps, PersonCard } from "./cards/PersonCard.tsx";
+import { InsightCardProps, InsightCard } from "./cards/InsightCard.tsx";
 import {
   TestimonialCardProps,
   TestimonialCard,
 } from "./cards/TestimonialCard.tsx";
+import {
+  StaticMapSection,
+  StaticMapSectionProps,
+} from "./pageSections/StaticMapSection.tsx";
 
 export interface PageSectionCategoryProps {
   BreadcrumbsSection: BreadcrumbsSectionProps;
   HeroSection: HeroSectionProps;
+  EventSection: EventSectionProps;
   CoreInfoSection: CoreInfoSectionProps;
+  InsightSection: InsightSectionProps;
   NearbyLocationsSection: NearbyLocationsSectionProps;
   BannerSection: BannerSectionProps;
+  ProductSection: ProductSectionProps;
   PromoSection: PromoSectionProps;
   PhotoGallerySection: PhotoGallerySectionProps;
-  FAQsSection: FAQsSectionProps;
-  SectionContainer: SectionContainerProps;
+  TeamSection: TeamSectionProps;
+  FAQSection: FAQSectionProps;
+  StaticMapSection: StaticMapSectionProps;
+  TestimonialSection: TestimonialSectionProps;
 }
 
 export const PageSectionCategoryComponents = {
@@ -73,15 +98,48 @@ export const PageSectionCategoryComponents = {
   CoreInfoSection,
   NearbyLocationsSection,
   BannerSection,
-  PromoSection,
   PhotoGallerySection,
-  FAQsSection,
-  SectionContainer,
+  StaticMapSection,
+  EventSection,
+  FAQSection,
+  InsightSection,
+  ProductSection,
+  PromoSection,
+  TeamSection,
+  TestimonialSection,
 };
 
 export const PageSectionCategory = Object.keys(
   PageSectionCategoryComponents
 ) as (keyof PageSectionCategoryProps)[];
+
+export interface OtherCategoryProps {
+  Header: HeaderProps;
+  Footer: FooterProps;
+}
+
+export const OtherCategoryComponents = {
+  Header,
+  Footer,
+};
+
+export const OtherCategory = Object.keys(
+  OtherCategoryComponents
+) as (keyof OtherCategoryProps)[];
+
+export interface DirectoryCategoryProps {
+  Directory: DirectoryProps;
+}
+
+export const DirectoryCategoryComponents = {
+  Directory,
+};
+
+export const DirectoryCategory = Object.keys(
+  DirectoryCategoryComponents
+) as (keyof DirectoryCategoryProps)[];
+
+/** THE CATEGORIES BELOW ARE NO LONGER SUPPORTED */
 
 export interface LayoutBlockCategoryProps {
   Collection: CollectionProps;
@@ -101,16 +159,16 @@ export const LayoutBlockCategory = Object.keys(
 
 export interface CardCategoryProps {
   EventCard: EventCardProps;
-  InsightCard: InsightCardProps;
   PersonCard: PersonCardProps;
+  InsightCard: InsightCardProps;
   ProductCard: ProductCardProps;
   TestimonialCard: TestimonialCardProps;
 }
 
 export const CardCategoryComponents = {
   EventCard,
-  InsightCard,
   PersonCard,
+  InsightCard,
   ProductCard,
   TestimonialCard,
 };
@@ -152,29 +210,3 @@ export const ContentBlockCategoryComponents = {
 export const ContentBlockCategory = Object.keys(
   ContentBlockCategoryComponents
 ) as (keyof ContentBlockCategoryProps)[];
-
-export interface OtherCategoryProps {
-  Header: HeaderProps;
-  Footer: FooterProps;
-}
-
-export const OtherCategoryComponents = {
-  Header,
-  Footer,
-};
-
-export const OtherCategory = Object.keys(
-  OtherCategoryComponents
-) as (keyof OtherCategoryProps)[];
-
-export interface DirectoryCategoryProps {
-  Directory: DirectoryProps;
-}
-
-export const DirectoryCategoryComponents = {
-  Directory,
-};
-
-export const DirectoryCategory = Object.keys(
-  DirectoryCategoryComponents
-) as (keyof DirectoryCategoryProps)[];