+++ starter/src/ve.config.tsx
@@ -5,16 +5,7 @@ import {
   PageSectionCategory,
   PageSectionCategoryComponents,
   PageSectionCategoryProps,
-  LayoutBlockCategory,
-  CardCategory,
-  ContentBlockCategory,
-  LayoutBlockCategoryComponents,
-  CardCategoryComponents,
-  ContentBlockCategoryComponents,
   OtherCategoryComponents,
-  LayoutBlockCategoryProps,
-  CardCategoryProps,
-  ContentBlockCategoryProps,
   OtherCategoryProps,
   DirectoryCategory,
   DirectoryCategoryComponents,
@@ -36,18 +27,12 @@ const LocatorCategory = Object.keys(
 
 interface MainProps
   extends PageSectionCategoryProps,
-    LayoutBlockCategoryProps,
-    CardCategoryProps,
-    ContentBlockCategoryProps,
     DirectoryCategoryProps,
     OtherCategoryProps,
     LocatorCategoryProps {}
 
 const components: Config<MainProps>["components"] = {
   ...PageSectionCategoryComponents,
-  ...LayoutBlockCategoryComponents,
-  ...CardCategoryComponents,
-  ...ContentBlockCategoryComponents,
   ...DirectoryCategoryComponents,
   ...OtherCategoryComponents,
   ...LocatorCategoryComponents,
@@ -60,18 +45,6 @@ export const mainConfig: Config<MainProps> = {
     pageSections: {
       title: "Page Sections",
       components: PageSectionCategory,
-    },
-    layoutBlocks: {
-      title: "Layout Blocks",
-      components: LayoutBlockCategory,
-    },
-    cardBlocks: {
-      title: "Cards",
-      components: CardCategory,
-    },
-    contentBlocks: {
-      title: "Content Blocks",
-      components: ContentBlockCategory,
     },
     directory: {
       title: "Directory",
@@ -87,11 +60,6 @@ export const mainConfig: Config<MainProps> = {
       return (
         <DropZone
           zone="default-zone"
-          disallow={[
-            ...ContentBlockCategory,
-            ...CardCategory,
-            ...LayoutBlockCategory,
-          ]}
           style={{ display: "flex", flexDirection: "column", height: "100vh" }}
         />
       );